requirements.txt
streamlit
pymilvus
sentence-transformers
pypdf
python-dotenv
#torch  # if using GPU